LW-21 Added settings namespace for mail_template_controller
