+++ app/views/settings/mail_templates/index.html.erb
